requirements.txt
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.3.0
requests==2.28.1